+++ src/main/drivers/rx_pwm.h
@@ -17,7 +17,7 @@
 
 #pragma once
 
-#include "drivers/io.h"
+#include "drivers/io_types.h"
 
 typedef enum {
     INPUT_FILTERING_DISABLED = 0,
