+++ src/main/fc/config.c
@@ -38,7 +38,7 @@
 #include "drivers/compass.h"
 #include "drivers/system.h"
 #include "drivers/timer.h"
-#include "drivers/pwm_rx.h"
+#include "drivers/rx_pwm.h"
 #include "drivers/rx_spi.h"
 #include "drivers/serial.h"
 #include "drivers/pwm_esc_detect.h"
