+++ src/main/main.c
@@ -45,7 +45,7 @@
 #include "drivers/accgyro.h"
 #include "drivers/compass.h"
 #include "drivers/pwm_esc_detect.h"
-#include "drivers/pwm_rx.h"
+#include "drivers/rx_pwm.h"
 #include "drivers/pwm_output.h"
 #include "drivers/adc.h"
 #include "drivers/bus_i2c.h"
