+++ src/main/rx/rx.c
@@ -31,7 +31,7 @@
 #include "config/feature.h"
 
 #include "drivers/adc.h"
-#include "drivers/pwm_rx.h"
+#include "drivers/rx_pwm.h"
 #include "drivers/rx_spi.h"
 #include "drivers/system.h"
